requirements.txt
gradio>=5.9.0
transformers>=4.36.0
huggingface-hub>=0.20.0
torch>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
pytest>=7.0.0
